Convert QA pipeline tasks - tickets/PIPE2D-1528 (#33)
* Create a `detectorMapQa.yaml` pipeline that contains two tasks, one for computing the DetectorMap residuals on a single exposure and the other for creating a aggregate statistics and plots for the entire collection.
* Clean up repository for lsst `gen3` transition.
  * Python 3.11
  * Removing unnecessary files
  * Fixing python namespacing
  * Reorganization of some of the files.
