--- bin.src/SConscript
@@ -1,3 +1,0 @@
-# -*- python -*-
-from lsst.sconsUtils import scripts
-scripts.BasicSConscript.shebang()
