--- bin.src/detectorMapQa.py
@@ -1,4 +1,0 @@
-#!/usr/bin/env python
-from pfs.drp.qa.detectorMapQa import DetectorMapQaTask
-
-DetectorMapQaTask.parseAndRun()
