--- bin.src/extractionQa.py
@@ -1,3 +1,0 @@
-#!/usr/bin/env python
-from pfs.drp.qa.extractionQa import ExtractionQaTask
-ExtractionQaTask.parseAndRun()
