--- bin.src/fluxCalQa.py
@@ -1,4 +1,0 @@
-#!/usr/bin/env python
-from pfs.drp.qa.fluxCalQa import FluxCalQaTask
-
-FluxCalQaTask.parseAndRun()
