+++ setup.cfg
@@ -10,5 +10,3 @@ exclude =
     python/pfs/drp/stella/utils/asinhNorm.py  # Imported from matplotlib, so different standards
 
 [tool:pytest]
-addopts = --flake8
-flake8-ignore = W503 E133 E203 E226 E228 N802 N803 N806 N812 N813 N815 N816 W504
